+++ src/main/java/edu/kit/tm/cm/scdm/sensordatastorage/application/dtos/service/DynamicVehicleDataDto.java
@@ -7,7 +7,7 @@ public interface DynamicVehicleDataDto {
 
     float getTirePressure();
 
-    float getTankLevel();
+    double getTankLevel();
 
     String getTimestamp();
 }
